Moved settings into a separate file
Changed README to refer to the settings file so that things are DRYer

Also: changed the description of the DISPLAY_UNIT field to be more useful
